+++ bot/src/handler.rs
@@ -1,4 +1,8 @@
-use crate::{bot::State, context::Context, utils::cp_from_container};
+use crate::{
+    bot::State,
+    context::Context,
+    utils::{cp_from_container, gen_pwd},
+};
 use anyhow::{Context as _, Result};
 use dptree::case;
 use log::{debug, info};
@@ -10,7 +14,11 @@ use teloxide::{
     macros::BotCommands,
     net::Download,
     prelude::*,
-    types::{InputFile, MediaKind, MessageKind, ReactionType, UpdateKind},
+    requests::HasPayload,
+    types::{
+        InputFile, MediaKind, MediaText, MessageCommon, MessageId, MessageKind, ReactionCount,
+        ReactionType, UpdateKind,
+    },
     utils::command::BotCommands as _,
 };
 use tokio::fs;
@@ -21,6 +29,8 @@ use tokio::fs;
     description = "These commands are supported:"
 )]
 enum Command {
+    #[command(description = "This is a bot to sync files from chat.")]
+    Start,
     #[command(description = "display this text.")]
     Help,
     #[command(description = "show the current state.")]
@@ -29,6 +39,8 @@ enum Command {
     Pause,
     #[command(description = "unpause the bot.")]
     Unpause,
+    #[command(description = "print current bypass password in the server side.")]
+    BypassPwd,
 }
 
 type MyDialogue = Dialogue<State, InMemStorage<State>>;
@@ -36,19 +48,30 @@ type MyDialogue = Dialogue<State, InMemStorage<State>>;
 pub fn handler() -> UpdateHandler<anyhow::Error> {
     dialogue::enter::<Update, InMemStorage<State>, State, _>()
         .branch(msg_handler())
-        .branch(react_handler())
+        .branch(channel_post_handler())
+        .branch(reaction_handler())
+        .branch(reaction_count_handler())
         .branch(callback_handler())
 }
 
 fn cmd_handler() -> UpdateHandler<anyhow::Error> {
     teloxide::filter_command::<Command, _>()
         .branch(
+            case![Command::Start].endpoint(async |bot: Bot, msg: Message| {{
+                bot.send_message(msg.chat.id, "This is a bot to sync files from chat. Enter /help to see all commands.")
+                    .await?;
+                Ok(())}})
+        )
+        .branch(
             case![Command::Help].endpoint(async |bot: Bot, msg: Message| {
                 bot.send_message(msg.chat.id, Command::descriptions().to_string())
                     .await?;
                 Ok(())
             }),
-        )
+        ).branch(case![Command::BypassPwd].endpoint(async |ctx: Context| {
+            info!("Bypass_pwd: {}", ctx.bypass_pwd.read().unwrap());
+            Ok(())
+        }))
         .branch(
             case![Command::State].endpoint(async |bot: Bot, msg: Message, state: State| {
                 bot.send_message(msg.chat.id, format!("{:?}", state))
@@ -58,11 +81,33 @@ fn cmd_handler() -> UpdateHandler<anyhow::Error> {
         )
         .branch(
             case![State::Paused].branch(case![Command::Unpause].endpoint(
-                async |bot: Bot, dialogue: MyDialogue, msg: Message, owner: UserId| {
-                    if msg.from.map(|user| user.id) != Some(owner) {
-                        bot.send_message(msg.chat.id, "Permission denied: You are not owner")
-                            .await?;
-                        dialogue.exit().await?;
+                async |bot: Bot, dialogue: MyDialogue, msg: Message, ctx: Context| {
+                    if msg.from.map(|user| ctx.bypass_users.contains(&user.id)) != Some(true) {
+                        // check bypass_pwd
+                        match msg {
+                            Message {
+                                kind:
+                                    MessageKind::Common(MessageCommon {
+                                        media_kind: MediaKind::Text(MediaText { text, .. }),
+                                        ..
+                                    }),
+                                ..
+                            } if text == format!("/unpause {}", ctx.bypass_pwd.read().unwrap()) => {
+                                // renew bypass_pwd
+                                let new = gen_pwd();
+                                info!("New bypass_pwd: {}", new);
+                                *ctx.bypass_pwd.write().unwrap() = new;
+                            }
+                            _ => {
+                                bot.send_message(
+                                    msg.chat.id,
+                                    "Permission denied: You are not in allow users list or invalid password",
+                                )
+                                .await?;
+                                dialogue.exit().await?;
+                                return Ok(());
+                            }
+                        }
                     }
                     dialogue.update(State::Working).await?;
                     bot.send_message(msg.chat.id, "Working").await?;
@@ -91,7 +136,7 @@ fn msg_handler() -> UpdateHandler<anyhow::Error> {
                         }
                         MediaKind::Document(document) => {
                             tokio::spawn(async move {
-                                handle_file(
+                                handle_msg_file(
                                     bot,
                                     ctx,
                                     &document.document.file.id,
@@ -108,7 +153,7 @@ fn msg_handler() -> UpdateHandler<anyhow::Error> {
                         }
                         MediaKind::Video(video) => {
                             tokio::spawn(async move {
-                                handle_file(
+                                handle_msg_file(
                                     bot,
                                     ctx,
                                     &video.video.file.id,
@@ -125,7 +170,7 @@ fn msg_handler() -> UpdateHandler<anyhow::Error> {
                                 if let Some(photo) =
                                     photo.photo.into_iter().max_by_key(|p| p.height)
                                 {
-                                    handle_file(
+                                    handle_msg_file(
                                         bot,
                                         ctx,
                                         &photo.file.id,
@@ -147,7 +192,7 @@ fn msg_handler() -> UpdateHandler<anyhow::Error> {
         .branch(dptree::endpoint(async || Ok(())))
 }
 
-fn react_handler() -> UpdateHandler<anyhow::Error> {
+fn reaction_handler() -> UpdateHandler<anyhow::Error> {
     Update::filter_message_reaction_updated().branch(case![State::Working].endpoint(
         async |bot: Bot, update: Update, ctx: Context| {
             if let UpdateKind::MessageReaction(reaction) = update.kind {
@@ -202,22 +247,146 @@ fn react_handler() -> UpdateHandler<anyhow::Error> {
     ))
 }
 
+fn reaction_count_handler() -> UpdateHandler<anyhow::Error> {
+    Update::filter_message_reaction_count_updated().branch(case![State::Working].endpoint(
+        async |bot: Bot, update: Update, ctx: Context| {
+            if let UpdateKind::MessageReactionCount(reaction) = update.kind {
+                let msg_id = reaction.message_id.0.to_ne_bytes();
+                let msgs = ctx
+                    .db
+                    .open_tree(reaction.chat.id.0.to_ne_bytes())
+                    .context("Failed to open msg tree")?;
+                if let Ok(Some(file_path)) = msgs.get(msg_id) {
+                    let file_path = std::path::Path::new(std::str::from_utf8(&file_path)?);
+                    let file_name = file_path
+                        .file_name()
+                        .and_then(|file| file.to_str())
+                        .context("Failed to read filename from db")?;
+
+                    let score: i32 = reaction
+                        .reactions
+                        .iter()
+                        .filter_map(|r| match r {
+                            ReactionCount {
+                                r#type: ReactionType::Emoji { emoji },
+                                total_count,
+                            } => match emoji.as_ref() {
+                                "👍" | "😁" | "🙏" | "😇" | "🤗" => {
+                                    Some(*total_count as i32)
+                                }
+                                "❤" | "🔥" | "🥰" | "🎉" | "🍌" | "💋" | "💘" | "😘" => {
+                                    Some(2 * *total_count as i32)
+                                }
+                                "❤‍🔥" => Some(3 * *total_count as i32),
+                                "👎" | "🤯" | "😱" | "😢" | "🥴" | "🌚" | "😐" | "🖕" | "😨" => {
+                                    Some(-(*total_count as i32))
+                                }
+                                "🤬" | "🤮" | "💩" | "🤡" | "💔" | "😡" => {
+                                    Some(-(2 * *total_count as i32))
+                                }
+                                _ => None,
+                            },
+                            _ => None,
+                        })
+                        .sum();
+
+                    if score >= ctx.fav_score_limit {
+                        let target_path = format!("{}/{}", ctx.fav_dir.display(), file_name);
+                        fs::rename(file_path, &target_path).await?;
+                        msgs.insert(msg_id, target_path.as_bytes())
+                            .context("Failed to update msg and file path")?;
+                        info!("Fav: {}", target_path);
+                    } else if score < ctx.delete_score_limit {
+                        msgs.remove(msg_id).context("Failed to remove msg_id")?;
+                        bot.delete_message(reaction.chat.id, reaction.message_id)
+                            .await?;
+                        fs::remove_file(&file_path).await?;
+                        info!("Delete: {}", file_path.display());
+                    }
+                }
+            }
+            Ok(())
+        },
+    ))
+}
+
+fn channel_post_handler() -> UpdateHandler<anyhow::Error> {
+    Update::filter_channel_post()
+        .branch(cmd_handler())
+        .branch(
+            case![State::Working].endpoint(async |bot: Bot, msg: Message, ctx: Context| {
+                if let MessageKind::Common(common_msg) = msg.kind {
+                    match common_msg.media_kind {
+                        MediaKind::Text(text) => {
+                            debug!("Text: {:#?}", text);
+                        }
+                        MediaKind::Document(document) => {
+                            tokio::spawn(async move {
+                                handle_channel_file(
+                                    bot,
+                                    ctx,
+                                    &document.document.file.id,
+                                    document
+                                        .document
+                                        .file_name
+                                        .unwrap_or(document.document.file.id.clone()),
+                                    msg.id,
+                                    msg.chat.id,
+                                )
+                                .await?;
+                                Result::<_, anyhow::Error>::Ok(())
+                            });
+                        }
+                        MediaKind::Video(video) => {
+                            tokio::spawn(async move {
+                                handle_channel_file(
+                                    bot,
+                                    ctx,
+                                    &video.video.file.id,
+                                    format!("{}.mp4", video.video.file.id),
+                                    msg.id,
+                                    msg.chat.id,
+                                )
+                                .await?;
+                                Result::<_, anyhow::Error>::Ok(())
+                            });
+                        }
+                        MediaKind::Photo(photo) => {
+                            tokio::spawn(async move {
+                                if let Some(photo) =
+                                    photo.photo.into_iter().max_by_key(|p| p.height)
+                                {
+                                    handle_channel_file(
+                                        bot,
+                                        ctx,
+                                        &photo.file.id,
+                                        format!("{}.jpg", photo.file.id),
+                                        msg.id,
+                                        msg.chat.id,
+                                    )
+                                    .await?;
+                                }
+                                Result::<_, anyhow::Error>::Ok(())
+                            });
+                        }
+                        _ => {}
+                    }
+                }
+                Ok(())
+            }),
+        )
+}
+
 fn callback_handler() -> UpdateHandler<anyhow::Error> {
     Update::filter_callback_query()
 }
 
-async fn handle_file<F, Fut>(
-    bot: Bot,
-    ctx: Context,
-    file_id: impl AsRef<str>,
-    file_name: impl AsRef<str>,
-    chat_id: ChatId,
-    reply: F,
-) -> Result<()>
-where
-    F: Fn(&Bot, ChatId, InputFile) -> Fut,
-    Fut: core::future::IntoFuture<Output = core::result::Result<Message, teloxide::RequestError>>,
-{
+async fn save_file(
+    bot: &Bot,
+    ctx: &Context,
+    file_id: &impl AsRef<str>,
+    file_name: &impl AsRef<str>,
+) -> Result<String> {
     info!("Saving: {}", file_id.as_ref());
     let server_path = loop {
         if let Ok(f) = bot.get_file(file_id.as_ref()).send().await {
@@ -231,11 +400,11 @@ where
             let mut file = fs::File::create(&save_path).await?;
             bot.download_file(&server_path, &mut file).await?;
         }
-        true => match ctx.container_manager {
+        true => match &ctx.container_manager {
             Some(container_manager) => {
                 cp_from_container(
                     container_manager,
-                    ctx.container_id.unwrap(),
+                    ctx.container_id.as_ref().unwrap(),
                     server_path,
                     &save_path,
                 )
@@ -246,7 +415,27 @@ where
             }
         },
     }
+    info!("Saved: {}", save_path);
+    Ok(save_path)
+}
 
+/// Handle file directly sent to the bot.
+/// Due to the limitation of the bot API, the msg sent directly to the bot will be
+/// prevented from being deleted or edited, so the bot forwards the file to the chat
+/// again which can be operated by the bot.
+async fn handle_msg_file<F, Fut>(
+    bot: Bot,
+    ctx: Context,
+    file_id: impl AsRef<str>,
+    file_name: impl AsRef<str>,
+    chat_id: ChatId,
+    reply: F,
+) -> Result<()>
+where
+    F: Fn(&Bot, ChatId, InputFile) -> Fut,
+    Fut: core::future::IntoFuture<Output = core::result::Result<Message, teloxide::RequestError>>,
+{
+    let save_path = save_file(&bot, &ctx, &file_id, &file_name).await?;
     let reply_id = reply(
         &bot,
         chat_id,
@@ -262,6 +451,33 @@ where
         .context("Failed to open msg tree")?;
     msgs.insert(reply_id, save_path.as_bytes())
         .context("Failed to save msg_id")?;
-    info!("Saved: {}", save_path);
+    Ok(())
+}
+
+async fn handle_channel_file(
+    bot: Bot,
+    ctx: Context,
+    file_id: impl AsRef<str>,
+    file_name: impl AsRef<str>,
+    message_id: MessageId,
+    chat_id: ChatId,
+) -> Result<()> {
+    let mut req = bot.set_message_reaction(chat_id, message_id);
+    req.payload_mut().reaction = Some(vec![ReactionType::Emoji {
+        emoji: "🫡".to_string(),
+    }]);
+    req.await?;
+    let save_path = save_file(&bot, &ctx, &file_id, &file_name).await?;
+    let mut req = bot.set_message_reaction(chat_id, message_id);
+    req.payload_mut().reaction = Some(vec![ReactionType::Emoji {
+        emoji: "👌".to_string(),
+    }]);
+    req.await?;
+    let msgs = ctx
+        .db
+        .open_tree(chat_id.0.to_ne_bytes())
+        .context("Failed to open msg tree")?;
+    msgs.insert(message_id.0.to_ne_bytes(), save_path.as_bytes())
+        .context("Failed to save msg_id")?;
     Ok(())
 }